Fix router events emitting
